+++ kernel/include/intr/isrs.h
@@ -18,12 +18,14 @@
 #ifndef _ISRS_H_
 #define _ISRS_H_
 
+#include <stddef.h>
+
 struct regs
 {
-    unsigned int gs, fs, es, ds;
-    unsigned int edi, esi, ebp, esp, ebx, edx, ecx, eax;
-    unsigned int int_no, err_code;
-    unsigned int eip, cs, eflags, useresp, ss;    
+    size_t gs, fs, es, ds;
+    size_t edi, esi, ebp, esp, ebx, edx, ecx, eax;
+    size_t int_no, err_code;
+    size_t eip, cs, eflags, useresp, ss;    
 };
 
 extern void isrs_install();
